# Conflicts:
#	src/Compent.uIntra/App_Plugins/Events/Details/DetailsView.cshtml
#	src/Compent.uIntra/App_Start/NinjectWebCommon.cs
#	src/Compent.uIntra/Compent.uIntra.csproj
#	src/Compent.uIntra/Core/PagePromotion/PagePromotionAutoMapperProfile.cs
#	src/Compent.uIntra/Views/Events/DetailsView.cshtml

+++ src/uIntra.Likes/Web/LikesControllerBase.cs
@@ -5,6 +5,7 @@ using System.Web.Mvc;
 using uIntra.Core;
 using uIntra.Core.Activity;
 using uIntra.Core.Extensions;
+using uIntra.Core.PagePromotion;
 using uIntra.Core.User;
 using Umbraco.Web;
 using Umbraco.Web.Mvc;
@@ -25,7 +26,8 @@ namespace uIntra.Likes.Web
             IActivitiesServiceFactory activitiesServiceFactory,
             IIntranetUserService<IIntranetUser> intranetUserService,
             ILikesService likesService,
-            IDocumentTypeAliasProvider documentTypeAliasProvider, UmbracoHelper umbracoHelper)
+            IDocumentTypeAliasProvider documentTypeAliasProvider,
+            UmbracoHelper umbracoHelper)
         {
             _activitiesServiceFactory = activitiesServiceFactory;
             _intranetUserService = intranetUserService;
@@ -37,7 +39,7 @@ namespace uIntra.Likes.Web
         public virtual PartialViewResult ContentLikes()
         {
             var guid = CurrentPage.GetGuidKey();
-            return Likes(_likesService.GetLikeModels(guid), guid);
+            return Likes(_likesService.GetLikeModels(guid), guid, showTitle: true);
         }
 
         public virtual PartialViewResult Likes(ILikeable likesInfo)
@@ -59,13 +61,20 @@ namespace uIntra.Likes.Web
                 return Likes(_likesService.GetLikeModels(model.CommentId.Value), model.ActivityId, model.CommentId);
             }
 
+            if (IsForPagePromotion(model))
+            {
+                var pagePromotionLikeInfo = AddActivityLike(model.ActivityId);
+                return Likes(pagePromotionLikeInfo.Likes, pagePromotionLikeInfo.Id, showTitle: true);
+            }
+
             if (IsForContentPage(model))
             {
                 _likesService.Add(GetCurrentUserId(), model.ActivityId);
                 return Likes(_likesService.GetLikeModels(model.ActivityId), model.ActivityId);
             }
 
-            return AddActivityLike(model.ActivityId);
+            var activityLikeInfo = AddActivityLike(model.ActivityId);
+            return Likes(activityLikeInfo.Likes, activityLikeInfo.Id);
         }
 
         [HttpPost]
@@ -77,13 +86,20 @@ namespace uIntra.Likes.Web
                 return Likes(_likesService.GetLikeModels(model.CommentId.Value), model.ActivityId, model.CommentId);
             }
 
+            if (IsForPagePromotion(model))
+            {
+                var pagePromotionLikeInfo = RemoveActivityLike(model.ActivityId);
+                return Likes(pagePromotionLikeInfo.Likes, pagePromotionLikeInfo.Id, showTitle: true);
+            }
+
             if (IsForContentPage(model))
             {
                 _likesService.Remove(GetCurrentUserId(), model.ActivityId);
                 return Likes(_likesService.GetLikeModels(model.ActivityId), model.ActivityId);
             }
 
-            return RemoveActivityLike(model.ActivityId);
+            var activityLikeInfo = RemoveActivityLike(model.ActivityId);
+            return Likes(activityLikeInfo.Likes, activityLikeInfo.Id);
         }
 
         protected virtual bool IsForComment(AddRemoveLikeModel model)
@@ -91,12 +107,18 @@ namespace uIntra.Likes.Web
             return model.CommentId.HasValue;
         }
 
+        protected virtual bool IsForPagePromotion(AddRemoveLikeModel model)
+        {
+            var content = _umbracoHelper.TypedContent(model.ActivityId);
+            return content != null && PagePromotionHelper.IsPagePromotion(content);
+        }
+
         protected virtual bool IsForContentPage(AddRemoveLikeModel model)
         {
             return _umbracoHelper.TypedContent(model.ActivityId)?.DocumentTypeAlias == _documentTypeAliasProvider.GetContentPage();
         }
 
-        protected virtual PartialViewResult Likes(IEnumerable<LikeModel> likes, Guid activityId, Guid? commentId = null, bool isReadOnly = false)
+        protected virtual PartialViewResult Likes(IEnumerable<LikeModel> likes, Guid activityId, Guid? commentId = null, bool isReadOnly = false, bool showTitle = false)
         {
             var currentUserId = GetCurrentUserId();
             var likeModels = likes as IList<LikeModel> ?? likes.ToList();
@@ -109,30 +131,24 @@ namespace uIntra.Likes.Web
                 Count = likeModels.Count,
                 CanAddLike = canAddLike,
                 Users = likeModels.Select(el => el.User),
-                IsReadOnly = isReadOnly
+                IsReadOnly = isReadOnly,
+                ShowTitle = showTitle
             };
             return PartialView(LikesViewPath, model);
         }
 
-        protected virtual PartialViewResult AddActivityLike(Guid activityId)
+        protected virtual Guid GetCurrentUserId() => _intranetUserService.GetCurrentUserId();
+
+        protected ILikeable AddActivityLike(Guid activityId)
         {
             var service = _activitiesServiceFactory.GetService<ILikeableService>(activityId);
-            var likeInfo = service.AddLike(GetCurrentUserId(), activityId);
-
-            return Likes(likeInfo.Likes, likeInfo.Id);
+            return service.AddLike(GetCurrentUserId(), activityId);
         }
 
-        protected virtual PartialViewResult RemoveActivityLike(Guid activityId)
+        protected ILikeable RemoveActivityLike(Guid activityId)
         {
             var service = _activitiesServiceFactory.GetService<ILikeableService>(activityId);
-            var likeInfo = service.RemoveLike(GetCurrentUserId(), activityId);
-
-            return Likes(likeInfo.Likes, likeInfo.Id);
-        }
-
-        protected virtual Guid GetCurrentUserId()
-        {
-            return _intranetUserService.GetCurrentUserId();
+            return service.RemoveLike(GetCurrentUserId(), activityId);
         }
     }
 }
